+++ lib.d.ts
@@ -1,3 +1,3 @@
-declare const _esm: any;
-declare function _require__esm(params?: {}): any;
-declare function _require__esm__cjs(): any;
+declare const esm: any;
+declare function esm_require(params?: {}): any;
+declare function esm_cjs_require(params?: {}): any;