+++ lib.ts
@@ -1,11 +1,13 @@
-const _esm = require('esm')
+const esm = require('esm')
 module.exports = {
-	_require__esm,
-	_require__esm__cjs,
+	esm_require,
+	_require__esm: esm_require,
+	esm_cjs_require,
+	_require__esm__cjs: esm_cjs_require,
 }
-function _require__esm(params={}) {
-  return _esm(module, params)
+function esm_require(params={}) {
+  return esm(module, params)
 }
-function _require__esm__cjs() {
-  return _require__esm({ cjs: true })
+function esm_cjs_require(params={}) {
+  return esm_require(params)
 }